[feature] Split HasBrowser trait
