+++ tests/HttpBrowserTest.php
@@ -4,13 +4,14 @@ namespace Zenstruck\Browser\Tests;
 
 use Symfony\Component\Panther\PantherTestCase;
 use Zenstruck\Browser\HttpBrowser;
+use Zenstruck\Browser\Test\HasHttpBrowser;
 
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
  */
 final class HttpBrowserTest extends PantherTestCase
 {
-    use BrowserKitBrowserTests;
+    use BrowserKitBrowserTests, HasHttpBrowser;
 
     /**
      * @test
