+++ tests/KernelBrowserTests.php
@@ -5,13 +5,14 @@ namespace Zenstruck\Browser\Tests;
 use Symfony\Bundle\FrameworkBundle\KernelBrowser as SymfonyKernelBrowser;
 use Symfony\Component\Security\Core\User\User;
 use Zenstruck\Browser\KernelBrowser;
+use Zenstruck\Browser\Test\HasBrowser;
 
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
  */
 trait KernelBrowserTests
 {
-    use BrowserKitBrowserTests;
+    use BrowserKitBrowserTests, HasBrowser;
 
     /**
      * @test
@@ -91,10 +92,5 @@ trait KernelBrowserTests
         ;
 
         $this->assertTrue($profile->hasCollector('request'));
-    }
-
-    protected function browser(): KernelBrowser
-    {
-        return $this->kernelBrowser();
     }
 }
