+++ tests/PantherBrowserTest.php
@@ -5,6 +5,7 @@ namespace Zenstruck\Browser\Tests;
 use PHPUnit\Framework\AssertionFailedError;
 use Symfony\Component\Panther\PantherTestCase;
 use Zenstruck\Browser\PantherBrowser;
+use Zenstruck\Browser\Test\HasPantherBrowser;
 
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
@@ -13,7 +14,7 @@ use Zenstruck\Browser\PantherBrowser;
  */
 final class PantherBrowserTest extends PantherTestCase
 {
-    use BrowserTests;
+    use BrowserTests, HasPantherBrowser;
 
     /**
      * @test
